Rename cmd/maintenanceUpdate to cmd/maintenance
Align the command package directory name with the CLI command name
("canasta maintenance") and the convention used by all other command
packages (e.g. cmd/extension, cmd/config, cmd/backup).
