+++ cmd/maintenance/extension_test.go
