+++ cmd/maintenance/maintenanceScript.go
